Modified code for u64

+++ src/lib.rs
@@ -1,7 +1,7 @@
 /// Copy of `pow()` from `std`.
 #[inline(never)]
-pub fn pow_std(mut base: u8, mut exp: u32) -> u8 {
-    let mut acc: u8 = 1;
+pub fn pow_std(mut base: u64, mut exp: u32) -> u64 {
+    let mut acc: u64 = 1;
 
     while exp > 1 {
         if (exp & 1) == 1 {
@@ -23,7 +23,7 @@ pub fn pow_std(mut base: u8, mut exp: u32) -> u8 {
 
 /// `pow()` from `std` with power-of-2 optimization.
 #[inline(never)]
-pub fn pow_std_2opt(mut base: u8, mut exp: u32) -> u8 {
+pub fn pow_std_2opt(mut base: u64, mut exp: u32) -> u64 {
     if base & base.wrapping_sub(1) == 0 {
         if base == 0 {
             if exp == 0 {
@@ -35,7 +35,7 @@ pub fn pow_std_2opt(mut base: u8, mut exp: u32) -> u8 {
         return 1 << (base_pow * exp);
     }
 
-    let mut acc: u8 = 1;
+    let mut acc: u64 = 1;
     while exp > 1 {
         if (exp & 1) == 1 {
             acc *= base;
@@ -56,8 +56,8 @@ pub fn pow_std_2opt(mut base: u8, mut exp: u32) -> u8 {
 
 /// Alternate `pow()` uninlined.
 #[inline(never)]
-pub fn pow_alt(mut base: u8, mut exp: u32) -> u8 {
-    let mut acc: u8 = 1;
+pub fn pow_alt(mut base: u64, mut exp: u32) -> u64 {
+    let mut acc: u64 = 1;
 
     loop {
         if (exp & 1) == 1 {
@@ -73,14 +73,14 @@ pub fn pow_alt(mut base: u8, mut exp: u32) -> u8 {
 
 /// Alternate `pow()` uninlined with 01-opt.
 #[inline(never)]
-pub fn pow_alt_01opt(mut base: u8, mut exp: u32) -> u8 {
+pub fn pow_alt_01opt(mut base: u64, mut exp: u32) -> u64 {
     if base <= 1 {
         if exp == 0 {
             return 1;
         }
         return base;
     }
-    let mut acc: u8 = 1;
+    let mut acc: u64 = 1;
     loop {
         if (exp & 1) == 1 {
             acc *= base;
@@ -95,7 +95,7 @@ pub fn pow_alt_01opt(mut base: u8, mut exp: u32) -> u8 {
 
 /// Alternate `pow()` uninlined with 01-opt and 2-opt.
 #[inline(never)]
-pub fn pow_alt_012opt(mut base: u8, mut exp: u32) -> u8 {
+pub fn pow_alt_012opt(mut base: u64, mut exp: u32) -> u64 {
     if base & base.wrapping_sub(1) == 0 {
         if base <= 1 {
             if exp == 0 {
@@ -107,7 +107,7 @@ pub fn pow_alt_012opt(mut base: u8, mut exp: u32) -> u8 {
         return 1 << (base_pow * exp);
     }
 
-    let mut acc: u8 = 1;
+    let mut acc: u64 = 1;
     loop {
         if (exp & 1) == 1 {
             acc *= base;
@@ -122,7 +122,7 @@ pub fn pow_alt_012opt(mut base: u8, mut exp: u32) -> u8 {
 
 /// Alternate `pow()` uninlined with 0-opt and 2-opt.
 #[inline(never)]
-pub fn pow_alt_02opt(mut base: u8, mut exp: u32) -> u8 {
+pub fn pow_alt_02opt(mut base: u64, mut exp: u32) -> u64 {
     if base & base.wrapping_sub(1) == 0 {
         if base == 0 {
             if exp == 0 {
@@ -134,7 +134,7 @@ pub fn pow_alt_02opt(mut base: u8, mut exp: u32) -> u8 {
         return 1 << (base_pow * exp);
     }
 
-    let mut acc: u8 = 1;
+    let mut acc: u64 = 1;
     loop {
         if (exp & 1) == 1 {
             acc *= base;
@@ -149,8 +149,8 @@ pub fn pow_alt_02opt(mut base: u8, mut exp: u32) -> u8 {
 
 /// Alternate `pow()` always inlined.
 #[inline(always)]
-pub fn pow_alt_inline(mut base: u8, mut exp: u32) -> u8 {
-    let mut acc: u8 = 1;
+pub fn pow_alt_inline(mut base: u64, mut exp: u32) -> u64 {
+    let mut acc: u64 = 1;
 
     loop {
         if (exp & 1) == 1 {
@@ -166,7 +166,7 @@ pub fn pow_alt_inline(mut base: u8, mut exp: u32) -> u8 {
 
 /// Alternate `pow()` uninlined with 2-opt.
 #[inline(never)]
-pub fn pow_alt_2opt(mut base: u8, mut exp: u32) -> u8 {
+pub fn pow_alt_2opt(mut base: u64, mut exp: u32) -> u64 {
     if base & base.wrapping_sub(1) == 0 {
         if base == 0 {
             if exp == 0 {
@@ -178,7 +178,7 @@ pub fn pow_alt_2opt(mut base: u8, mut exp: u32) -> u8 {
         return 1 << (base_pow * exp);
     }
 
-    let mut acc: u8 = 1;
+    let mut acc: u64 = 1;
     loop {
         if (exp & 1) == 1 {
             acc *= base;
@@ -193,7 +193,7 @@ pub fn pow_alt_2opt(mut base: u8, mut exp: u32) -> u8 {
 
 /// Alternate `pow()` always inlined with 2-opt.
 #[inline(always)]
-pub fn pow_alt_2opt_inline(mut base: u8, mut exp: u32) -> u8 {
+pub fn pow_alt_2opt_inline(mut base: u64, mut exp: u32) -> u64 {
     if base & base.wrapping_sub(1) == 0 {
         if base == 0 {
             if exp == 0 {
@@ -205,7 +205,7 @@ pub fn pow_alt_2opt_inline(mut base: u8, mut exp: u32) -> u8 {
         return 1 << (base_pow * exp);
     }
 
-    let mut acc: u8 = 1;
+    let mut acc: u64 = 1;
     loop {
         if (exp & 1) == 1 {
             acc *= base;
